% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{append_typecons}
\alias{append_typecons}
\title{Append TYPECONS}
\usage{
append_typecons(dt)
}
\arguments{
\item{dt}{A data.table containing required fields.}
}
\value{
A data.table with TYPECONS and TYCONS2.
}
\description{
Append TYPECONS, TYCONS2 compo at the end of input table.
}
