% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{buff_and_remove_streets_batch}
\alias{buff_and_remove_streets_batch}
\title{Add buffer and remove streets polygons batch optimized version (> 200 polygons).}
\usage{
buff_and_remove_streets_batch(polys, streets)
}
\arguments{
\item{polys}{An sf object of POLYGON.}

\item{streets}{An sf object of LINESTRING with LANE_RADIUS.}
}
\value{
An sf object
}
\description{
This function remove streets from an sf table of POINTS polygons.
}
