% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{calculate_radius}
\alias{calculate_radius}
\title{Calculate radius}
\usage{
calculate_radius(dt)
}
\arguments{
\item{dt}{A data.table containing required fields.}
}
\value{
A data.table with an added output column.
}
\description{
This function calculates the radius for a risk in order to create neighbors risks pockets
later on. It is based on many risk characteristics.
}
