% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{get_joint_risks}
\alias{get_joint_risks}
\title{Polygon blaster}
\usage{
get_joint_risks(dt)
}
\arguments{
\item{dt}{A data.frame with the following}
}
\value{
PRCH_ID and POLYGON_ID with package version.
}
\description{
This function return polygon id from an input data.table using
the precomputed pockets.
}
\examples{
\dontrun{
dt <- jsonlite::fromJSON('[{"PRCH_ID":14543671,"COMAUBAT":"NA",
"AFFECTAT":"C8112","LATITCOM":"45.6388","LONGICOM":"-73.8438",
"PRINCFUS":4,"RISASGRB":"1","SUPERREZ":1800,
"UMESSUP2":"PI","TYPECONS":5,"TYCONS2":"NA","RISKRADIUS":18.2958}]')
get_polygons_id(dt)
}
}
