% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{get_risks_cgen}
\alias{get_risks_cgen}
\title{Get risks CGEN from extraw}
\usage{
get_risks_cgen()
}
\value{
A data.table object with IDs and columns.
}
\description{
Query to extract inforce commercial policies details for cgen to use in joint risks evaluation.
}
