% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{get_risks_ugen}
\alias{get_risks_ugen}
\title{Get risks UGEN from extraw}
\usage{
get_risks_ugen()
}
\value{
A data.table object with IDs and columns.
}
\description{
Query to extract inforce commercial policies details for ugen to use in joint risks evaluation.
}
