% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{load_risk_ugen}
\alias{load_risk_ugen}
\title{Get risks UGEN from a file}
\usage{
load_risk_ugen(file)
}
\arguments{
\item{file}{file to load}
}
\value{
A data.table object with IDs and columns.
}
\description{
Query to extract inforce commercial policies details for ugen to use in joint risks evaluation.
}
