% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{set_binding}
\alias{set_binding}
\title{Change the loaded in memory polygons}
\usage{
set_binding(binding, value)
}
\arguments{
\item{pockets}{A character string, name of netezza catalog.}
}
\description{
Change the loaded in memory polygons
}
