% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{update_polygons}
\alias{update_polygons}
\title{Obtain polygons}
\usage{
update_polygons(source)
}
\arguments{
\item{source}{A data.table used to update the polygon. Use load_risk_cgen(file) or
get_cgen_risk() (require extractnetezza).}
}
\value{
A data.table with PRCH_ID and POLYGON_ID.
}
\description{
This function is used to update and return polygon table.
}
