% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{append_polygons_idx}
\alias{append_polygons_idx}
\title{Add polygon index}
\usage{
append_polygons_idx(
  dt,
  prefix = "",
  affectat = paste0(prefix, "AFFECTAT"),
  comaubat = paste0(prefix, "COMAUBAT"),
  latitcom = paste0(prefix, "LATITCOM"),
  longicom = paste0(prefix, "LONGICOM"),
  murstruc = paste0(prefix, "MURSTRUC"),
  plancher = paste0(prefix, "PLANCHER"),
  pregeoco = paste0(prefix, "PREGEOCO"),
  princfus = paste0(prefix, "PRINCFUS"),
  resaufeu = paste0(prefix, "RESAUFEU"),
  risasgrb = paste0(prefix, "RISASGRB"),
  rvextbet = paste0(prefix, "RVEXTBET"),
  rvextbri = paste0(prefix, "RVEXTBRI"),
  superrez = paste0(prefix, "SUPERREZ"),
  toitstru = paste0(prefix, "TOITSTRU"),
  umessupe = paste0(prefix, "UMESSUPE"),
  mttotras = paste0(prefix, "MTTOTRAS"),
  streets
)
}
\arguments{
\item{dt}{A data.table to which you want to add polygon index. If the field name
in the table does not match the default value, just change the parameter for
that field. Field value are equivalence value (COOP_ID 11).}

\item{prefix}{A character string. In case the fields just need a prefix (something like "PROD_ID"). Default to "".}

\item{affectat}{A character string. Default to paste0(prefix, "AFFECTAT").}

\item{comaubat}{A character string. Default to paste0(prefix, "COMAUBAT").}

\item{latitcom}{A character string. Default to paste0(prefix, "LATITCOM").}

\item{longicom}{A character string. Default to paste0(prefix, "LONGICOM").}

\item{murstruc}{A character string. Default to paste0(prefix, "MURSTRUC").}

\item{plancher}{A character string. Default to paste0(prefix, "PLANCHER").}

\item{pregeoco}{A character string. Default to paste0(prefix, "PREGEOCO").}

\item{princfus}{A character string. Default to paste0(prefix, "PRINCFUS").}

\item{resaufeu}{A character string. Default to paste0(prefix, "RESAUFEU").}

\item{risasgrb}{A character string. Default to paste0(prefix, "RISASGRB").}

\item{rvextbet}{A character string. Default to paste0(prefix, "RVEXTBET").}

\item{rvextbri}{A character string. Default to paste0(prefix, "RVEXTBRI").}

\item{superrez}{A character string. Default to paste0(prefix, "SUPERREZ").}

\item{toitstru}{A character string. Default to paste0(prefix, "TOITSTRU").}

\item{umessupe}{A character string. Default to paste0(prefix, "UMESSUPE").}

\item{mttotras}{A character string. Default to paste0(prefix, "MTTOTRAS").}

\item{streets}{A LINESTRING object of navstreets with LANE_RADIUS and geometry elements.}
}
\value{
A data.table with paste0(prefix,"POLYINDX"), paste0(prefix,"POLYMAXRISASGRB"),
paste0(prefix,"POLYSUMMTTOTRAS") columns. If the risk has no appropriate geolocation
information, index will be NA and other values will just be the individual risk values.
}
\description{
Adds a polygon index to a data.table of commercial risks.
Use the others parameters for custom input names but all of .
}
\examples{
\dontrun{
dt <- extraw::get_policies(
    inforce = TRUE,
    partial = "PRO",
    filters = list(MPROD_ID = c(2552251, 1071124, 1071125, 1071122),
                   MLIAF_ID = 4,
                   MCAAF_ID = 2),
    detailid = c(140, 959, 971, 1082, 1083, 9045, 9406, 9408, 14218, 14219, 14220, 14367, 14491, 14650, 14660, 14661)
  )
append_polygons_idx(dt, prefix = "PROD_", comaubat = "PROD_14367")
}
}
