% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{create_polygons}
\alias{create_polygons}
\title{Polygon-o-tron}
\usage{
create_polygons(dt)
}
\arguments{
\item{dt}{A data.table containing required fields.}
}
\value{
A data.table.
}
\description{
This function takes a data containing IDs, radius and geocode and creates pockets
for neighbors risks. It returns a data containing the pocket number for each PRCH_ID
}
