% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{get_joint_risks}
\alias{get_joint_risks}
\title{Polygon blaster}
\usage{
get_joint_risks(dt, polygons, streets)
}
\arguments{
\item{dt}{A data.frame with the following}

\item{polygons}{A data.frame to compare dt to.}

\item{streets}{A LINESTRING object of navstreets with LANE_RADIUS and geometry elements.}
}
\value{
PRCH_ID and POLYGON_ID with package version.
}
\description{
This function return polygon id from an input data.table using
the precomputed pockets.
}
\examples{
\dontrun{
dt <- jsonlite::fromJSON('[{"PRCH_ID":14543671,"COMAUBAT":"NA",
"AFFECTAT":"C8112","LATITCOM":"46.77418", "LONGICOM":"-71.30196",
"PRINCFUS":4,"SUPERREZ":1800, "UMESSUPE":"PI","TYPECONS":5}]')
get_joint_risks(dt, polygons)
}
}
