% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointrisk.R
\name{update_polygons}
\alias{update_polygons}
\title{Obtain polygons}
\usage{
update_polygons(source, polygons = NULL, streets = NULL)
}
\arguments{
\item{source}{A data.table used to update the polygon. Use load_risk_cgen(file) or
get_cgen_risk() (require extraw).}

\item{polygons}{A polygons ref to update from.}

\item{streets}{A LINESTRING object of navstreets with LANE_RADIUS and geometry elements.}
}
\value{
A list with a message and a spatial feature data frame.
}
\description{
This function is used to update and return polygon table.
}
